+++ app/(tabs)/goal-settings.tsx
@@ -1,0 +1,31 @@
+import { StyleSheet } from 'react-native';
+import { ThemedView } from '@/components/ThemedView';
+import { ThemedText } from '@/components/ThemedText';
+
+export default function GoalSettingsScreen() {
+  return (
+    <ThemedView style={styles.container}>
+      <ThemedText type="title" style={styles.title}>
+        目標設定
+      </ThemedText>
+      <ThemedText style={styles.description}>
+        日、週、月、年ごとの目標を設定できます。
+      </ThemedText>
+    </ThemedView>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 16,
+  },
+  title: {
+    fontSize: 28,
+    marginBottom: 16,
+  },
+  description: {
+    fontSize: 16,
+    marginBottom: 24,
+  },
+});
