+++ app/(tabs)/statistics.tsx
@@ -1,0 +1,31 @@
+import { StyleSheet } from 'react-native';
+import { ThemedView } from '@/components/ThemedView';
+import { ThemedText } from '@/components/ThemedText';
+
+export default function StatisticsScreen() {
+  return (
+    <ThemedView style={styles.container}>
+      <ThemedText type="title" style={styles.title}>
+        統計データ
+      </ThemedText>
+      <ThemedText style={styles.description}>
+        日、週、月、年ごとのグラフをタブで切り替えて表示します。
+      </ThemedText>
+    </ThemedView>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 16,
+  },
+  title: {
+    fontSize: 28,
+    marginBottom: 16,
+  },
+  description: {
+    fontSize: 16,
+    marginBottom: 24,
+  },
+});
